get process needs to be updated implemented for unix
