+++ go.mod
@@ -13,37 +13,27 @@ require (
 	github.com/buger/jsonparser v0.0.0-20180910192245-6acdf747ae99
 	github.com/client9/misspell v0.3.4 // indirect
 	github.com/fatih/gomodifytags v0.0.0-20180914191908-141225bf62b6 // indirect
-	github.com/fsnotify/fsnotify v1.4.7
 	github.com/gladiusio/gladius-utils v0.0.0-20180827165816-9ab431d232e7
 	github.com/golang/lint v0.0.0-20181023015148-c2de83ec1785 // indirect
 	github.com/google/shlex v0.0.0-20150127133951-6f45313302b9 // indirect
 	github.com/gordonklaus/ineffassign v0.0.0-20180909121442-1003c8bd00dc // indirect
-	github.com/gorilla/context v1.1.1
+	github.com/gorilla/context v1.1.1 // indirect
 	github.com/gorilla/mux v1.6.2
 	github.com/gorilla/websocket v1.4.0
-	github.com/hashicorp/errwrap v1.0.0
 	github.com/hashicorp/go-multierror v1.0.0
-	github.com/hashicorp/hcl v1.0.0
 	github.com/jgautheron/goconst v0.0.0-20170703170152-9740945f5dcb // indirect
-	github.com/kardianos/osext v0.0.0-20170510131534-ae77be60afb1
-	github.com/kardianos/service v0.0.0-20180910224244-b1866cf76903
+	github.com/kardianos/osext v0.0.0-20170510131534-ae77be60afb1 // indirect
+	github.com/kardianos/service v0.0.0-20180910224244-b1866cf76903 // indirect
 	github.com/kisielk/errcheck v1.1.0 // indirect
-	github.com/magiconair/properties v1.8.0
 	github.com/mdempsky/gocode v0.0.0-20180727200127-00e7f5ac290a // indirect
 	github.com/mdempsky/maligned v0.0.0-20180708014732-6e39bd26a8c8 // indirect
 	github.com/mdempsky/unconvert v0.0.0-20180703203632-1a9a0a0a3594 // indirect
 	github.com/mibk/dupl v1.0.0 // indirect
-	github.com/mitchellh/mapstructure v1.0.0
 	github.com/nicksnyder/go-i18n v1.10.0 // indirect
 	github.com/opennota/check v0.0.0-20180911053232-0c771f5545ff // indirect
-	github.com/pelletier/go-toml v1.2.0
 	github.com/rogpeppe/godef v1.0.0 // indirect
 	github.com/securego/gosec v0.0.0-20181011124531-3116b07de424 // indirect
 	github.com/sirupsen/logrus v1.0.6
-	github.com/spf13/afero v1.1.2
-	github.com/spf13/cast v1.2.0
-	github.com/spf13/jwalterweatherman v1.0.0
-	github.com/spf13/pflag v1.0.2
 	github.com/spf13/viper v1.2.0
 	github.com/sqs/goreturns v0.0.0-20180302073349-83e02874ec12 // indirect
 	github.com/stripe/safesql v0.0.0-20171221195208-cddf355596fe // indirect
@@ -52,12 +42,10 @@ require (
 	github.com/walle/lll v0.0.0-20160702150637-8b13b3fbf731 // indirect
 	github.com/zmb3/goaddimport v0.0.0-20170810013102-4ab94a07ab86 // indirect
 	github.com/zmb3/gogetdoc v0.0.0-20180522031303-10095872a7c5 // indirect
-	golang.org/x/crypto v0.0.0-20180910181607-0e37d006457b
+	golang.org/x/crypto v0.0.0-20180910181607-0e37d006457b // indirect
 	golang.org/x/lint v0.0.0-20181023015148-c2de83ec1785 // indirect
-	golang.org/x/sys v0.0.0-20181026144532-2772b66316d2
-	golang.org/x/text v0.3.0
+	golang.org/x/sys v0.0.0-20181026203630-95b1ffbd15a5
 	gopkg.in/alecthomas/kingpin.v3-unstable v3.0.0-20180810215634-df19058c872c // indirect
-	gopkg.in/yaml.v2 v2.2.1
 	honnef.co/go/tools v0.0.0-20180920025451-e3ad64cb4ed3 // indirect
 	mvdan.cc/interfacer v0.0.0-20180901003855-c20040233aed // indirect
 	mvdan.cc/unparam v0.0.0-20181021212735-8e9846b49e8a // indirect
